pjon/pjdl: add capture of minimal PJON frames over PJDL link layer
The PJON protocol can use several link layers. These captures run on
PJDL, demonstrate a minimal PJON frame (with TX info) and a response.
And also contain a few exceptional conditions: Glitches in the signal,
outside and inside of frames. Lack of recipient's response, too.

The snippets were extracted from a capture that was
Submitted-By: Julio Aguirre <jcallano@gmail.com>