caliper: rename dump files and probes
Rename captures to the "caliper*.sr" pattern. This avoids names which
start with dashes and could result in surprises. Users still can tell
whether to expect positive or negative values without too much effort.

Rename the probes. It's absolutely confusing to provide RXD and TXD
names for a protocol that has CLK and DATA signals. And users had to
manually assign the traces to decoder inputs after guessing which might
be which (it's not as obvious for calipers as I2C or SPI would be).