sle44xx: extend the README for the SLE4442 captures
Create a README in the sle44xx/ parent directory, discuss the chip
family's features, and reference the Siemens datasheet. Extend the
README for the SLE4442 captures, don't assume that all users are
familiar with the chips and their protocol. The example dumps also
serve for learning and exploration of the curious. Fixup the probe
names (according to --show their names are in the 0..7 range).