Makefile: use PREFIX in "make install", use DESTDIR in traditional ways
Commit 4fafe8159940 as of 2012-05-14 introduced a simple "make install"
target. Unfortunately it used the DESTDIR variable to mean the location
where captures get installed to (relative to the target filesystem's
root). But traditionally DESTDIR is an optional "root directory" owned
by packagers or used for cross build installs. While the install location
within the filesystem should be relative to a --prefix spec which users
can control during build.

Use DESTDIR during installation in its traditional meaning, accept when
it's specified externally, but don't modify it locally. Use PREFIX to
mean the in-filesystem base directory which share/... gets appended to.
Provide a backwards compatible builtin, accept an external override.

  $ make PREFIX=${HOME} install
  $ make DESTDIR=$( pwd )/rootfs install

Add a developer comment. Ideally we would not have to roll DIY shell
commands, and use a higher level build tool instead. The reason that
this has not been done yet might be the portable construction of the
list of directories to install at compile time (a guess).